Refactor password and fix bug

+++ src/password/password.service.ts
@@ -1,13 +1,28 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 import { Password, Prisma } from '@prisma/client'
 
 @Injectable()
 export class PasswordService {
-  constructor(private prisma: PrismaService) { }
+  constructor(private readonly prisma: PrismaService) { }
 
-  async create(data: Prisma.PasswordCreateInput): Promise<Password> {
-    return this.prisma.password.create({
+  async create(data: Prisma.PasswordCreateInput) {
+    // Query the database to check if url and username is already in use.
+    const urlUsernameUserIdExists = await this.prisma.password.findUnique({
+      where: {
+        url_username_userId: {
+          url: String(data.url),
+          username: String(data.username),
+          userId: Number(data.user.connect?.id)
+        }
+      }
+    })
+  
+    if (urlUsernameUserIdExists) {
+      throw new ConflictException('Login already exists');
+    }
+
+    return await this.prisma.password.create({
       data,
     });
   }
@@ -30,20 +45,16 @@ export class PasswordService {
   }
 
   async findOne(passwordWhereUniqueInput: Prisma.PasswordWhereUniqueInput): Promise<Password | null> {
-    return this.prisma.password.findUnique({
+    const password = await this.prisma.password.findUnique({
       where: passwordWhereUniqueInput
     });
-  }
 
-  async findUnique(passwordWhereInput: Prisma.PasswordWhereInput): Promise<Password | null> {
-    return this.prisma.password.findFirst({
-      where: {
-        AND: [
-          { url: passwordWhereInput.url },
-          { username: passwordWhereInput.username }
-        ]
-      }
-    })
+    // Check if result is null
+    if (!password) {
+      throw new NotFoundException('Password not found');
+    }
+
+    return password;
   }
 
   async update(params: {
@@ -51,15 +62,39 @@ export class PasswordService {
     data: Prisma.PasswordUpdateInput;
   }): Promise<Password> {
     const { where, data } = params;
-    return this.prisma.password.update({
+    // Query the database to check if a password with that id exists.
+    const passwordExists = await this.findOne({ id: where.id });
+
+    // Check if passwordExists is null
+    if (!passwordExists) {
+      throw new NotFoundException('Password not found');
+    }
+    const password = await this.prisma.password.update({
       where,
       data,
     });
+
+    // Check if the constant passwordExists is null.
+    if (!password) {
+      throw new NotFoundException('Password not Found');
+    }
+
+    return password;
   }
 
   async remove(where: Prisma.PasswordWhereUniqueInput): Promise<Password> {
-    return this.prisma.password.delete({
+    // Query the database to check if a password with that id exists.
+    const passwordExists = await this.findOne({ id: where.id });
+
+    // Check if the constant passwordExists is null.
+    if (!passwordExists) {
+      throw new NotFoundException('Password not Found');
+    }
+
+    const password = await this.prisma.password.delete({
       where,
     });
+
+    return password;
   }
 }
